[Standup] Standardize the use non-default service accounts all steps
A new CLI option, `-q/--serviceaccount` is available to specify service
accounts

An example script to create service accounts with fine-grained
permissions is made available
