+++ scenarios/examples/cpu.sh
@@ -82,6 +82,19 @@ cat << EOF > ${LLMDBENCH_VLLM_COMMON_EXTRA_VOLUMES}
     secretName: ${LLMDBENCH_VLLM_COMMON_CONTEXT_SECRET_NAME}
 EOF
 
+#export LLMDBENCH_VLLM_COMMON_EXTRA_CONTAINER_CONFIG=$(mktemp)
+#cat << EOF > ${LLMDBENCH_VLLM_COMMON_EXTRA_CONTAINER_CONFIG}
+#securityContext:
+#  capabilities:
+#    add:
+#    - IPC_LOCK
+#    - SYS_RAWIO
+#    - NET_ADMIN
+#  runAsGroup: 0
+#  runAsUser: 0
+#imagePullPolicy: Always
+#EOF
+
 export LLMDBENCH_VLLM_COMMON_PREPROCESS="python3 /setup/preprocess/set_llmdbench_environment.py && . \$HOME/llmdbench_env.sh"
 
 export LLMDBENCH_VLLM_STANDALONE_ARGS=$(mktemp)
@@ -121,6 +134,8 @@ export LLMDBENCH_VLLM_MODELSERVICE_DECODE_POD_LABELS=$LLMDBENCH_VLLM_COMMON_POD_
 export LLMDBENCH_VLLM_MODELSERVICE_DECODE_EXTRA_VOLUME_MOUNTS=$LLMDBENCH_VLLM_COMMON_EXTRA_VOLUME_MOUNTS
 export LLMDBENCH_VLLM_MODELSERVICE_DECODE_EXTRA_VOLUMES=$LLMDBENCH_VLLM_COMMON_EXTRA_VOLUMES
 export LLMDBENCH_VLLM_MODELSERVICE_DECODE_PREPROCESS=$LLMDBENCH_VLLM_COMMON_PREPROCESS
+#export LLMDBENCH_VLLM_MODELSERVICE_DECODE_EXTRA_CONTAINER_CONFIG=${LLMDBENCH_VLLM_COMMON_EXTRA_CONTAINER_CONFIG}
+
 export LLMDBENCH_VLLM_MODELSERVICE_DECODE_MODEL_COMMAND=custom
 export LLMDBENCH_VLLM_MODELSERVICE_DECODE_EXTRA_ARGS=$(mktemp)
 cat << EOF > $LLMDBENCH_VLLM_MODELSERVICE_DECODE_EXTRA_ARGS
